reorganizing
doing this so that we can use the methods from this pkg in other pkgs
(such as tests in the snitch-wasm repo).
