+++ hostfunc/httprequest.go
@@ -1,4 +1,4 @@
-package snitch
+package hostfunc
 
 import (
 	"bytes"
@@ -10,65 +10,15 @@ import (
 	"github.com/pkg/errors"
 	"github.com/streamdal/snitch-protos/build/go/protos/steps"
 	"github.com/tetratelabs/wazero/api"
+
+	"github.com/streamdal/snitch-go-client/helper"
 )
 
-// HostFuncKVExists is function that is exported to and called from a Rust WASM module
-func (s *Snitch) HostFuncKVExists(_ context.Context, module api.Module, ptr, length int32) int32 {
-	// Read request
-	request := &steps.KVExistsRequest{}
-
-	if err := ReadRequestFromMemory(module, request, ptr, length); err != nil {
-		return kvExistsResponse(module, "unable to read KVExistsRequest params: "+err.Error(), false, false)
-	}
-
-	if err := validateKVExistsRequest(request); err != nil {
-		return kvExistsResponse(module, "unable to validate KVExistsRequest: "+err.Error(), false, false)
-	}
-
-	// Perform operation
-	exists := s.kv.Exists(request.Key)
-
-	// Return response
-	return kvExistsResponse(module, "lookup succeeded", false, exists)
-}
-
-// Generates a protobuf response, writes to mem, and returns ptr to mem
-func kvExistsResponse(module api.Module, msg string, isError, exists bool) int32 {
-	resp := &steps.KVExistsResponse{
-		Exists:  exists,
-		IsError: isError,
-		Message: msg,
-	}
-
-	addr, err := WriteResponseToMemory(module, resp)
-	if err != nil {
-		panic("unable to write KVExistsResponse to memory: " + err.Error())
-	}
-
-	return addr
-}
-
-func validateKVExistsRequest(request *steps.KVExistsRequest) error {
-	if request == nil {
-		return errors.New("request cannot be nil")
-	}
-
-	if request.Key == "" {
-		return errors.New("request.Key cannot be empty")
-	}
-
-	if request.Mode != steps.KVExistsMode_KV_EXISTS_MODE_STATIC && request.Mode != steps.KVExistsMode_KV_EXISTS_MODE_DYNAMIC {
-		return errors.New("request.Mode must be either KV_EXISTS_MODE_STATIC or KV_EXISTS_MODE_DYNAMIC")
-	}
-
-	return nil
-}
-
-// HostFuncHTTPRequest is function that is exported to and called from a Rust WASM module
-func HostFuncHTTPRequest(_ context.Context, module api.Module, ptr, length int32) int32 {
+// HTTPRequest is function that is exported to and called from a Rust WASM module
+func (h *HostFunc) HTTPRequest(_ context.Context, module api.Module, ptr, length int32) int32 {
 	request := &steps.HttpRequest{}
 
-	if err := ReadRequestFromMemory(module, request, ptr, length); err != nil {
+	if err := helper.ReadRequestFromMemory(module, request, ptr, length); err != nil {
 		return httpRequestResponse(module, http.StatusInternalServerError, "unable to read HTTP request params: "+err.Error(), nil)
 	}
 
@@ -115,7 +65,7 @@ func httpRequestResponse(module api.Module, code int, body string, headers map[s
 		Headers: headers,
 	}
 
-	addr, err := WriteResponseToMemory(module, resp)
+	addr, err := helper.WriteResponseToMemory(module, resp)
 	if err != nil {
 		panic("unable to write HTTP response to memory: " + err.Error())
 	}
