+++ snitch.go
@@ -28,6 +28,7 @@ import (
 
 	"github.com/streamdal/snitch-protos/build/go/protos"
 
+	"github.com/streamdal/snitch-go-client/hostfunc"
 	"github.com/streamdal/snitch-go-client/kv"
 	"github.com/streamdal/snitch-go-client/logger"
 	"github.com/streamdal/snitch-go-client/metrics"
@@ -91,6 +92,7 @@ type Snitch struct {
 	audiencesMtx       *sync.RWMutex
 	sessionID          string
 	kv                 kv.IKV
+	hf                 *hostfunc.HostFunc
 	tailsMtx           *sync.RWMutex
 	tails              map[string]map[string]*Tail
 }
@@ -161,6 +163,11 @@ func New(cfg *Config) (*Snitch, error) {
 		return nil, errors.Wrap(err, "failed to start kv service")
 	}
 
+	hf, err := hostfunc.New(kvInstance, cfg.Logger)
+	if err != nil {
+		return nil, errors.Wrap(err, "failed to create hostfunc instance")
+	}
+
 	s := &Snitch{
 		functions:          make(map[string]*function),
 		functionsMtx:       &sync.RWMutex{},
@@ -175,6 +182,7 @@ func New(cfg *Config) (*Snitch, error) {
 		metrics:            m,
 		sessionID:          uuid.New().String(),
 		kv:                 kvInstance,
+		hf:                 hf,
 		tailsMtx:           &sync.RWMutex{},
 		tails:              make(map[string]map[string]*Tail),
 	}
